requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
build==1.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chromadb==1.0.20
click==8.2.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
e2b==1.5.1
e2b-code-interpreter==1.5.1
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.19.1
firecrawl-py==2.8.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.40.3
google-search-results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.9
hpack==4.1.0
html-to-markdown==1.10.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json-repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-mongodb==0.6.2
langchain-neo4j==0.4.0
langchain-ollama==0.3.7
langchain-openai==0.3.23
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.11
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-checkpoint-mongodb==0.1.4
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lark==1.2.2
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
motor==3.7.1
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
narwhals==2.5.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
numpy==2.3.0
oauthlib==3.3.1
ollama==0.5.3
onnxruntime==1.22.1
openai==1.86.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pillow==11.2.1
portalocker==2.10.1
posthog==5.4.0
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pygments==2.19.1
pymongo==4.12.1
pypdf==5.6.0
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.3
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.41
starlette==0.46.2
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typer==0.16.0
types-pyyaml==6.0.12.20250516
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
